// src/types/Synergetic/Finance/iSynGeneralLedagerJournal.ts
type iSynGeneralLedgerJournal =  {
  GLJournalSeq: number;
  GLYear: number;
  GLCode: string;
  GLSuballocationCode: string;
  PostingNumber: number;
  GLDate: Date;
  PostingSource: string;
  ReceiptNumber: number;
  DebtorTransactionSeq: number;
  CreditorTransactionSeq: number;
  GLAmount: number;
  JournalDescription: string;
  ReverseGLJournalSeq: number;
  PayCode: string;
  SaleInvoiceNumber: number;
  TaxCode: string;
  TaxAmount: number;
  IncomeFlag: boolean;
  InvoiceNumberIssued: number;
  ID: number;
  GLClassificationCode: string;
  AdjustmentFlag: boolean;
  BASPostingNumber: number;
  InvoiceLineDescription: string | null;
  OrderNumber: number;
  DebtorTransactionXRefSeq: number | null;
  GeneralLedgerJournalCreatedDate: Date | null;
  GeneralLedgerJournalCreatedByID: number | null;
  DebtorTransactionSeqCharge: number;
  ControlJournalFlag: boolean;
  TaxIncomeControlGLSeq: number | null;
  TaxExpenseControlGLSeq: number | null;
  ForeignCurrencyCode: string;
  ForeignCurrencyExchangeRate: number | null;
  ForeignCurrencyGLAmount: number;
  ForeignCurrencyTaxAmount: number;
  ModifiedDate: Date | null;
  ModifiedByID: number | null;
  AllowBankTransactionFlag: boolean;
  BankFeeFlag: boolean;
  StockMovementGLJournalSeq: number | null;
  StockControlGLJournalSeq: number | null;
  PostingNumberTransfer: number | null;
  PrintVersion: number | null;
  ReplicateSourceSynDatabaseCode: string | null;
  ReplicateSourcePostingNumber: number | null;
  ReplicateSourceGLJournalSeq: number | null;
};

export default iSynGeneralLedgerJournal;